# lib/api-zod/src/generated/types/entityInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shradha Oil Center ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { EntityInputPricingTier } from './entityInputPricingTier';
import type { EntityInputType } from './entityInputType';

export interface EntityInput {
  type: EntityInputType;
  name: string;
  mobile: string;
  gstin?: string;
  address?: string;
  city?: string;
  state?: string;
  pricingTier?: EntityInputPricingTier;
  creditLimit?: number;
}